# specs/009-embeddable-react-widget/contracts/widget-api.openapi.yaml
openapi: 3.1.0
info:
  title: Maintainer's Copilot Widget API
  version: 0.1.0
  description: Phase 9 embeddable widget, loader, public config, and widget chat API.
paths:
  /admin/widget-configs:
    get:
      summary: List widget configurations
      operationId: listWidgetConfigs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin-visible widget configurations.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WidgetConfigRead"
        "403":
          $ref: "#/components/responses/AdminRequired"
    post:
      summary: Create widget configuration
      operationId: createWidgetConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetConfigCreate"
      responses:
        "201":
          description: Widget configuration created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetConfigRead"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "422":
          $ref: "#/components/responses/InvalidWidgetConfig"
  /admin/widget-configs/{config_id}:
    patch:
      summary: Update widget configuration
      operationId: updateWidgetConfig
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetConfigUpdate"
      responses:
        "200":
          description: Widget configuration updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetConfigRead"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "404":
          $ref: "#/components/responses/WidgetNotFound"
        "422":
          $ref: "#/components/responses/InvalidWidgetConfig"
    delete:
      summary: Delete or disable a widget configuration
      operationId: deleteWidgetConfig
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      responses:
        "204":
          description: Widget configuration deleted or disabled and audited.
        "403":
          $ref: "#/components/responses/AdminRequired"
        "404":
          $ref: "#/components/responses/WidgetNotFound"
  /admin/widget-configs/{config_id}/embed-snippet:
    get:
      summary: Get one-script widget embed snippet
      operationId: getWidgetEmbedSnippet
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConfigId"
      responses:
        "200":
          description: Widget embed snippet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedSnippetRead"
        "403":
          $ref: "#/components/responses/AdminRequired"
        "404":
          $ref: "#/components/responses/WidgetNotFound"
  /public/widgets/{widget_id}/config:
    get:
      summary: Read public widget configuration
      operationId: readPublicWidgetConfig
      parameters:
        - $ref: "#/components/parameters/WidgetId"
        - name: origin
          in: query
          required: false
          schema:
            type: string
          description: Optional host origin declared by the loader or iframe. The backend validates observed request origin or referrer first and treats this value as advisory only.
      responses:
        "200":
          description: Public runtime widget configuration.
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicWidgetConfig"
        "403":
          $ref: "#/components/responses/OriginBlocked"
        "404":
          $ref: "#/components/responses/WidgetNotFound"
        "409":
          $ref: "#/components/responses/WidgetDisabled"
  /public/widgets/{widget_id}/session:
    post:
      summary: Issue widget-scoped anonymous session token
      operationId: issueWidgetSession
      parameters:
        - $ref: "#/components/parameters/WidgetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetSessionRequest"
      responses:
        "201":
          description: Widget-scoped anonymous session token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetSessionToken"
        "403":
          $ref: "#/components/responses/OriginBlocked"
        "404":
          $ref: "#/components/responses/WidgetNotFound"
        "409":
          $ref: "#/components/responses/WidgetDisabled"
  /public/widgets/{widget_id}/chat/messages:
    post:
      summary: Submit widget chat message
      operationId: submitWidgetChatMessage
      parameters:
        - $ref: "#/components/parameters/WidgetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WidgetChatSubmitRequest"
      responses:
        "202":
          description: Widget chat message accepted for streaming.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetChatAccepted"
        "403":
          $ref: "#/components/responses/OriginBlocked"
        "404":
          $ref: "#/components/responses/WidgetNotFound"
        "409":
          $ref: "#/components/responses/WidgetDisabled"
  /public/widgets/{widget_id}/chat/stream:
    get:
      summary: Stream public widget chat response
      operationId: streamWidgetChat
      parameters:
        - $ref: "#/components/parameters/WidgetId"
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: conversation_id
          in: query
          required: true
          schema:
            type: string
        - name: stream_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Streaming widget chat response.
          headers:
            X-Request-ID:
              schema:
                type: string
            X-Trace-ID:
              schema:
                type: string
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/WidgetChatStreamEvent"
        "403":
          $ref: "#/components/responses/OriginBlocked"
        "404":
          $ref: "#/components/responses/WidgetNotFound"
        "409":
          $ref: "#/components/responses/WidgetDisabled"
        "422":
          description: Invalid chat input or invalid widget session token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /widget/loader.js:
    get:
      summary: Serve widget loader script
      operationId: getWidgetLoader
      responses:
        "200":
          description: JavaScript loader script.
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/javascript:
              schema:
                type: string
  /widget/frame/{widget_id}:
    get:
      summary: Serve iframe widget shell
      operationId: getWidgetFrame
      parameters:
        - $ref: "#/components/parameters/WidgetId"
        - name: origin
          in: query
          required: false
          schema:
            type: string
          description: Optional host origin declared by the loader. The backend validates observed request origin or referrer first and treats this value as advisory only.
      responses:
        "200":
          description: Widget iframe shell.
          headers:
            Content-Security-Policy:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
          content:
            text/html:
              schema:
                type: string
        "403":
          description: Origin blocked.
        "404":
          description: Widget not found.
        "409":
          description: Widget disabled.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ConfigId:
      name: config_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    WidgetId:
      name: widget_id
      in: path
      required: true
      schema:
        type: string
        minLength: 8
  responses:
    AdminRequired:
      description: Admin role required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidWidgetConfig:
      description: Invalid widget configuration.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    WidgetNotFound:
      description: Widget configuration not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    WidgetDisabled:
      description: Widget is disabled.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    OriginBlocked:
      description: Origin is not allowed for the widget.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    WidgetConfigCreate:
      type: object
      required: [allowed_origins, theme, greeting, enabled_tools, enabled]
      properties:
        allowed_origins:
          type: array
          minItems: 1
          items:
            type: string
        theme:
          type: string
          enum: [light, dark, system]
        greeting:
          type: string
          minLength: 1
          maxLength: 500
        position:
          type: string
          enum: [bottom-right, bottom-left]
          default: bottom-right
        enabled_tools:
          type: array
          items:
            type: string
            enum: [classifier, ner, summarizer, rag]
        enabled:
          type: boolean
    WidgetConfigUpdate:
      type: object
      minProperties: 1
      properties:
        allowed_origins:
          type: array
          minItems: 1
          items:
            type: string
        theme:
          type: string
          enum: [light, dark, system]
        greeting:
          type: string
          minLength: 1
          maxLength: 500
        position:
          type: string
          enum: [bottom-right, bottom-left]
        enabled_tools:
          type: array
          items:
            type: string
            enum: [classifier, ner, summarizer, rag]
        enabled:
          type: boolean
    WidgetConfigRead:
      type: object
      required:
        - widget_id
        - allowed_origins
        - theme
        - greeting
        - enabled_tools
        - enabled
        - created_by
        - created_at
        - updated_at
      properties:
        widget_id:
          type: string
        allowed_origins:
          type: array
          items:
            type: string
        theme:
          type: string
        greeting:
          type: string
        position:
          type: string
        enabled_tools:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EmbedSnippetRead:
      type: object
      required: [widget_id, snippet]
      properties:
        widget_id:
          type: string
        snippet:
          type: string
        generated_at:
          type: string
          format: date-time
    PublicWidgetConfig:
      type: object
      required:
        [widget_id, theme, greeting, position, enabled_tools, session_endpoint, message_submit_endpoint, chat_stream_endpoint]
      properties:
        widget_id:
          type: string
        theme:
          type: string
        greeting:
          type: string
        position:
          type: string
        enabled_tools:
          type: array
          items:
            type: string
        session_endpoint:
          type: string
        message_submit_endpoint:
          type: string
        chat_stream_endpoint:
          type: string
        asset_base_url:
          type: string
    WidgetSessionRequest:
      type: object
      required: [origin]
      properties:
        origin:
          type: string
          description: Client-declared host origin for cross-checking. The backend must not trust this as the sole authorization signal.
    WidgetSessionToken:
      type: object
      required: [widget_id, token, conversation_id, expires_at]
      properties:
        widget_id:
          type: string
        token:
          type: string
        conversation_id:
          type: string
        expires_at:
          type: string
          format: date-time
    WidgetChatSubmitRequest:
      type: object
      required: [token, conversation_id, message]
      properties:
        token:
          type: string
        conversation_id:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
    WidgetChatAccepted:
      type: object
      required: [widget_id, conversation_id, stream_id, stream_url]
      properties:
        widget_id:
          type: string
        conversation_id:
          type: string
        stream_id:
          type: string
        stream_url:
          type: string
    WidgetChatStreamEvent:
      type: object
      required: [event_type, sequence]
      properties:
        event_type:
          type: string
          enum: [message_delta, warning, error, done]
        sequence:
          type: integer
          minimum: 0
        content:
          type: string
        request_id:
          type: string
        trace_id:
          type: string
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - authentication_required
            - admin_required
            - invalid_widget_config
            - widget_not_found
            - widget_disabled
            - origin_not_allowed
            - invalid_chat_input
            - chat_unavailable
            - internal_error
        message:
          type: string
        request_id:
          type: string
        trace_id:
          type: string
        details:
          type: object
          additionalProperties: true